% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/py_to_r_df.R
\name{convert_filtered_to_df}
\alias{convert_filtered_to_df}
\title{Helper function to convert (ideally, filtered) py_df
to a data.frame in R.}
\usage{
convert_filtered_to_df(py_obj, dtype)
}
\arguments{
\item{py_obj}{Reference to a python object (pandas.DataFrame) to convert}
}
\value{
R data.frame
}
\description{
Converts object to a matrix -> data.frame & then applies
column names to the data.frame
}
\details{
Note: does not handle any type-conversion; types are
cast to their lowest-denominator type, which might
be that of a python object
}
