% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/py_cohort_to_r_df.R
\name{py_cohort_to_r_df}
\alias{py_cohort_to_r_df}
\title{Convenience function to convert a (python) Cohort
into an R data.frame for use in further analysis}
\usage{
py_cohort_to_r_df(py_cohort, dtypes = c("inexact", "bool_", "object",
  "integer"), ...)
}
\arguments{
\item{py_cohort}{Python object representing a Cohort}

\item{dtypes}{which dtypes to convert}

\item{...}{(other parameters to `as_dataframe()`}
}
\description{
Uses \code{reticulate} to interface with python libraries.
Depends on having `cohorts` installed in local python environment.
}
