% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/py_to_r_df.R
\name{py_to_r_df}
\alias{py_to_r_df}
\title{Convert pandas.DataFrame to an R DataFrame}
\usage{
py_to_r_df(py_df, dtypes = ALLOWED_DTYPES)
}
\arguments{
\item{py_df}{Python object (pandas.DataFrame) to be converted}

\item{dtypes}{which dtypes to convert, as character vector}
}
\value{
R data.frame
}
\description{
Note: not a perfect conversion! Does not handle dates, lists, etc.
Not likely to work except for the simplest dataframes. For each dtype,
converts pandas DataFrame to an R data.frame via a matrix.
}
